Fixed MRRMSM type

+++ src/enzo/global_data.h
@@ -270,7 +270,7 @@ EXTERN FLOAT MultiRefineRegionRadius[MAX_STATIC_REGIONS+MAX_TRACKS];
 EXTERN FLOAT MultiRefineRegionWidth[MAX_STATIC_REGIONS+MAX_TRACKS];
 EXTERN int MultiRefineRegionMaximumLevel[MAX_STATIC_REGIONS+MAX_TRACKS];
 EXTERN int MultiRefineRegionMinimumLevel[MAX_STATIC_REGIONS+MAX_TRACKS];
-EXTERN int MultiRefineRegionMinimumStarMass[MAX_STATIC_REGIONS+MAX_TRACKS];
+EXTERN FLOAT MultiRefineRegionMinimumStarMass[MAX_STATIC_REGIONS+MAX_TRACKS];
 EXTERN int MultiRefineRegionMaximumOuterLevel;
 EXTERN int MultiRefineRegionMinimumOuterLevel;
 EXTERN int MultiRefineRegionSpatiallyVaryingStarMass;
